Splits BeingStatus into CharacterStatus and CritterStatus.
Should have been like this from the start but I was lazy  ^_^U
